+++ docs/Changelog.md
@@ -4,6 +4,7 @@
 -------------------------
 * [#4561](https://github.com/TouK/nussknacker/pull/4561) Improvement: Users are not requrired at OAuth2 config file
 * [#4492](https://github.com/TouK/nussknacker/pull/4492) Allow to test fragments using ad hoc testing. 
+* [#4575](https://github.com/TouK/nussknacker/pull/4575) Feature: Flink Kafka cacheable producer
 
 1.11.0 (Not released yet)
 -------------------------